refactor: Remove double/float shim overloads and redundant casts (#151)
* remove double overload for MakeSingleAction

* remove all unnecessary .0

* remove .0f

* remove MakeScale downcast

* remove MakeColor downcast

* remove MakeRotation downcast

* use direct float casting

* use direct float casting pt. 2

* use direct float casting pt. 3

* remove MakePos downcast

* remove AddAnimationDelayLoopFirstLast downcast

* remove unnecessary cast

* remove more cast shims

* remove unnecessary casting

* remove Ceil downcast

* remove Round downcast

* remove downcast for Cosf, Sinf Tanf, Acosf

* use float directly

* remove redundant float castings

* add override keyword

* remove unnecessary double cast

* use MathF

* use mathF pt. 2

* use MathF pt. 3

+++ CutTheRope/Commons/CtrRenderer.cs
@@ -235,7 +235,7 @@ namespace CutTheRope.Commons
             SCREEN_BG_SCALE_X = SCREEN_WIDTH_EXPANDED / SCREEN_WIDTH;
             if (IS_WVGA)
             {
-                SCREEN_WIDE_BG_SCALE_Y = (float)(SCREEN_HEIGHT_EXPANDED * 1.5 / 800.0);
+                SCREEN_WIDE_BG_SCALE_Y = SCREEN_HEIGHT_EXPANDED * 1.5f / 800;
                 SCREEN_WIDE_BG_SCALE_X = SCREEN_BG_SCALE_X;
                 return;
             }
